sensor: generate flush_complete event when data timestamps matches
Generating a flush_complete event right after receiving
 a flush_request from the sensor services breaks the flow
 as the client stops acquiring data.

Solution is to store the timestamp of the flush request
 and issue the flush_complete event when the timestamp
 of the sensor data matches

+++ libsensors_iio/src/Pressure.cpp
@@ -26,7 +26,7 @@ Pressure::Pressure(HWSensorBaseCommonData *data, const char *name,
 		sensor_t_data.flags |= SENSOR_FLAG_WAKE_UP;
 
 	sensor_t_data.resolution = data->channels[0].scale;
-	sensor_t_data.maxRange = sensor_t_data.resolution * (pow(2, data->channels[0].bits_used) - 1);
+	sensor_t_data.maxRange = sensor_t_data.resolution * (pow(2.0, (double)data->channels[0].bits_used) - 1);
 
 	num_data_axis = SENSOR_BASE_1AXIS;
 }
